Revert "Digits for renewable clusters : data series off the common series class and consequences"
This reverts commit 2a306ab8ee12355b71d8a19ae448f4ebb435f7d7.
