+++ src/libs/antares/study/parts/renewable/cluster.h
@@ -32,7 +32,6 @@
 #include "../../../array/matrix.h"
 #include "defines.h"
 #include "../common/cluster.h"
-#include "series.h"
 #include "../../fwd.h"
 
 namespace Antares
@@ -159,16 +158,10 @@ public:
     //@}
 
     bool setTimeSeriesModeFromString(const YString& value);
-    int saveDataSeriesToFolder(const AnyString& folder) const;
-    int loadDataSeriesFromFolder(Study& s, const AnyString& folder);
-
 
     const YString& getTimeSeriesModeAsString() const;
 
 public:
-    //! Series
-    DataSeriesRenewableCluster* series;
-
     /*!
     ** \brief The group ID
     **
