+++ src/libs/antares/study/parts/renewable/cluster_list.h
@@ -23,27 +23,6 @@ public:
     void estimateMemoryUsage(StudyMemoryUsage&) const override;
     bool loadFromFolder(const AnyString& folder, Area* area);
     bool saveToFolder(const AnyString& folder) const override;
-
-    /*!
-    ** \brief Resize all matrices dedicated to the sampled timeseries numbers
-    **
-    ** \param n A number of years
-    */
-    void resizeAllTimeseriesNumbers(uint n);
-
-    bool storeTimeseriesNumbers(Study& study);
-
-    int loadDataSeriesFromFolder(Study& study,
-        const StudyLoadOptions& options,
-        const AnyString& folder,
-        bool fast);
-
-    int saveDataSeriesToFolder(const AnyString& folder) const;
-
-    int saveDataSeriesToFolder(const AnyString& folder, const YString& msg) const;
-
-    void ensureDataTimeSeries();
-
 }; // class RenewableClusterList
 } // namespace Data
 } // namespace Antares
