+++ src/libs/antares/study/parts/thermal/cluster.cpp
@@ -78,7 +78,6 @@ namespace Data
 {
 Data::ThermalCluster::ThermalCluster(Area* parent, uint nbParallelYears) :
  Cluster(parent),
- series(nullptr),
  groupID(thermalDispatchGrpOther),
  mustrun(false),
  mustrunOrigin(false),
@@ -123,7 +122,6 @@ Data::ThermalCluster::ThermalCluster(Area* parent, uint nbParallelYears) :
 
 Data::ThermalCluster::ThermalCluster(Area* parent) :
  Cluster(parent),
- series(nullptr),
  groupID(thermalDispatchGrpOther),
  mustrun(false),
  mustrunOrigin(false),
@@ -258,7 +256,7 @@ void Data::ThermalCluster::copyFrom(const ThermalCluster& cluster)
     if (not prepro)
         prepro = new PreproThermal();
     if (not series)
-        series = new DataSeriesThermalCluster();
+        series = new DataSeriesCommon();
 
     prepro->copyFrom(*cluster.prepro);
     // timseries
@@ -425,15 +423,7 @@ void Data::ThermalCluster::reset()
     if (productionCost)
         (void)::memset(productionCost, 0, HOURS_PER_YEAR * sizeof(double));
 
-    unitCount = 0;
-    enabled = true;
-    nominalCapacity = 0.;
-
-    if (not series)
-        series = new DataSeriesThermalCluster();
-
-    series->series.reset(1, HOURS_PER_YEAR);
-    series->series.flush();
+    Cluster::reset();
 
     mustrun = false;
     mustrunOrigin = false;
@@ -690,47 +680,6 @@ bool ThermalCluster::checkMinStablePowerWithNewModulation(uint index, double val
     }
 
     return checkMinStablePower();
-}
-
-int ThermalCluster::saveDataSeriesToFolder(const AnyString& folder) const
-{
-    if (not folder.empty())
-    {
-        Yuni::Clob buffer;
-
-        buffer.clear() << folder << SEP << parentArea->id << SEP << id();
-        if (Yuni::IO::Directory::Create(buffer))
-        {
-            int ret = 1;
-            buffer.clear() << folder << SEP << parentArea->id << SEP << id() << SEP << "series.txt";
-            ret = series->series.saveToCSVFile(buffer, 0) && ret;
-
-            return ret;
-        }
-        return 0;
-    }
-    return 1;
-}
-
-int ThermalCluster::loadDataSeriesFromFolder(Study& s, const AnyString& folder)
-{
-    if (not folder.empty())
-    {
-        auto& buffer = s.bufferLoadingTS;
-
-        int ret = 1;
-        buffer.clear() << folder << SEP << parentArea->id << SEP << id() << SEP << "series."
-            << s.inputExtension;
-        ret = series->series.loadFromCSVFile(buffer, 1, HOURS_PER_YEAR, &s.dataBuffer) && ret;
-
-        if (s.usedByTheSolver && s.parameters.derated)
-            series->series.averageTimeseries();
-
-        series->timeseriesNumbers.clear();
-
-        return ret;
-    }
-    return 1;
 }
 
 } // namespace Data
