+++ src/libs/antares/study/parts/thermal/cluster_list.cpp
@@ -2,6 +2,18 @@
 #include "cluster.h"
 #include "../../study.h"
 
+namespace // anonymous
+{
+struct TSNumbersPredicate
+{
+    Yuni::uint32 operator()(Yuni::uint32 value) const
+    {
+        return value + 1;
+    }
+};
+
+} // anonymous namespace
+
 namespace Antares
 {
 namespace Data
@@ -31,7 +43,7 @@ void ThermalClusterList::estimateMemoryUsage(StudyMemoryUsage& u) const
         cluster.modulation.estimateMemoryUsage(u, true, thermalModulationMax, HOURS_PER_YEAR);
 
         if (cluster.series)
-            cluster.series->estimateMemoryUsage(u);
+            cluster.series->estimateMemoryUsage(u, timeSeriesThermal);
         if (cluster.prepro)
             cluster.prepro->estimateMemoryUsage(u);
 
@@ -698,111 +710,6 @@ void ThermalClusterList::enableMustrunForEveryone()
 {
     // enabling the mustrun mode
     each([&](ThermalCluster& cluster) { cluster.mustrun = true; });
-}
-
-void ThermalClusterList::resizeAllTimeseriesNumbers(uint n)
-{
-    assert(n < 200000); // arbitrary number
-    if (not cluster.empty())
-    {
-        if (0 == n)
-        {
-            each([&](ThermalCluster& cluster) { cluster.series->timeseriesNumbers.clear(); });
-        }
-        else
-        {
-            each([&](ThermalCluster& cluster) { cluster.series->timeseriesNumbers.resize(1, n); });
-        }
-    }
-}
-
-bool ThermalClusterList::storeTimeseriesNumbers(Study& study)
-{
-    if (cluster.empty())
-        return true;
-
-    bool ret = true;
-    TSNumbersPredicate predicate;
-
-    each([&](const ThermalCluster& cluster) {
-        study.buffer = study.folderOutput;
-        study.buffer << SEP << "ts-numbers" << SEP << typeID() << SEP << cluster.parentArea->id
-            << SEP << cluster.id() << ".txt";
-        ret = cluster.series->timeseriesNumbers.saveToCSVFile(study.buffer, 0, true, predicate)
-            and ret;
-    });
-    return ret;
-}
-
-int ThermalClusterList::saveDataSeriesToFolder(const AnyString& folder) const
-{
-    if (empty())
-        return 1;
-
-    int ret = 1;
-
-    auto end = cluster.end();
-    for (auto it = cluster.begin(); it != end; ++it)
-    {
-        auto& cluster = *(it->second);
-        if (cluster.series)
-            ret = cluster.saveDataSeriesToFolder(folder) and ret;
-    }
-    return ret;
-}
-
-int ThermalClusterList::saveDataSeriesToFolder(const AnyString& folder, const String& msg) const
-{
-    if (empty())
-        return 1;
-
-    int ret = 1;
-    uint ticks = 0;
-
-    auto end = cluster.end();
-    for (auto it = cluster.begin(); it != end; ++it)
-    {
-        auto& cluster = *(it->second);
-        if (cluster.series)
-        {
-            logs.info() << msg << "  " << (ticks * 100 / (1 + this->cluster.size()))
-                << "% complete";
-            ret = cluster.saveDataSeriesToFolder(folder) and ret;
-        }
-        ++ticks;
-    }
-    return ret;
-}
-
-int ThermalClusterList::loadDataSeriesFromFolder(Study& s,
-    const StudyLoadOptions& options,
-    const AnyString& folder,
-    bool fast)
-{
-    if (empty())
-        return 1;
-
-    int ret = 1;
-
-    each([&](ThermalCluster& cluster) {
-        if (cluster.series and (!fast or !cluster.prepro))
-            ret = cluster.loadDataSeriesFromFolder(s, folder) and ret;
-
-        ++options.progressTicks;
-        options.pushProgressLogs();
-    });
-    return ret;
-}
-
-void ThermalClusterList::ensureDataTimeSeries()
-{
-    auto end = cluster.end();
-    for (auto it = cluster.begin(); it != end; ++it)
-    {
-        auto& cluster = *(it->second);
-        if (not cluster.series)
-            cluster.series = new DataSeriesThermalCluster();
-    }
 }
 
 void ThermalClusterList::ensureDataPrepro()
