+++ src/libs/antares/study/parts/thermal/cluster_list.h
@@ -63,26 +63,6 @@ public:
 
     void estimateMemoryUsage(StudyMemoryUsage&) const override;
 
-    /*!
-    ** \brief Resize all matrices dedicated to the sampled timeseries numbers
-    **
-    ** \param n A number of years
-    */
-    void resizeAllTimeseriesNumbers(uint n);
-
-    bool storeTimeseriesNumbers(Study& study);
-
-    int loadDataSeriesFromFolder(Study& study,
-        const StudyLoadOptions& options,
-        const AnyString& folder,
-        bool fast);
-
-    int saveDataSeriesToFolder(const AnyString& folder) const;
-
-    int saveDataSeriesToFolder(const AnyString& folder, const YString& msg) const;
-
-    void ensureDataTimeSeries();
-
     Map mapping;
 
     /*!
